newsily-backend/requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
requests==2.26.0